requirements.txt
# Geometry Dash AI dependencies

neat-python>=0.92
opencv-python>=4.8.0
mss>=9.0.1
pyautogui>=0.9.54
numpy>=1.24.0
Pillow>=10.0.0

pywin32>=306          ; sys_platform == "win32"

pygetwindow>=0.0.9

keyboard>=0.13.5

